refactor: rename useToggle.ts to use-toggle.ts and update related imports
